requirements.txt
Django==1.10.6
PyYAML==3.12
Unidecode==0.04.20
Werkzeug==0.12
argcomplete==1.8.2
argparse==1.2.1
base58==0.2.4
boto3==1.4.4
botocore==1.5.26
click==6.7
coverage==4.3.4
django-allauth==0.31.0
django-annoying==0.10.3
django-extensions==1.7.7
django-material==0.13.1
django-organizations==0.9.1
django-picklefield==0.3.2
django-scheduler==0.8.3
docutils==0.13.1
durationpy==0.3
futures==3.0.5
hjson==2.0.2
httplib2==0.10.3
icalendar==3.11.3
jmespath==0.9.2
kappa==0.6.0
lambda-packages==0.14.2
oauthlib==2.0.2
placebo==0.8.1
python-dateutil==2.6.0
python-openid==2.2.5
python-slugify==1.2.1
pytz==2016.10
requests==2.13.0
requests-oauthlib==0.8.0
s3transfer==0.1.10
six==1.10.0
toml==0.9.2
tqdm==4.11.2
troposphere==1.9.2
twilio==5.7.0
wheel==0.29.0
wsgi-request-logger==0.4.6
wsgiref==0.1.2
zappa==0.39.1
